static_cast qsizetype index to int
QStringList::indexOf() returns an int in Qt 5 but a qsizetype in Qt 6.
Cast the index to int in order to let compilers deduce std::max's
template parameter type while building against Qt 6.
QComboBox::setCurrentIndex() still takes an int in Qt 6. The number of
supported Meson backends is not going to be out of the 32-bit integer
range.

+++ plugins/meson/settings/mesonadvancedsettings.cpp
@@ -38,7 +38,7 @@ void MesonAdvancedSettings::setConfig(const MesonAdvancedSettings::Data& conf)
 {
     m_ui->i_mesonArgs->setText(conf.args);
     m_ui->i_mesonExe->setUrl(conf.meson.toUrl());
-    m_ui->i_backend->setCurrentIndex(std::max(0, m_backendList.indexOf(conf.backend)));
+    m_ui->i_backend->setCurrentIndex(std::max(0, static_cast<int>(m_backendList.indexOf(conf.backend))));
 }
 
 void MesonAdvancedSettings::setSupportedBackends(const QStringList& backends)
